requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
streamlit==1.44.0
streamlit-autorefresh==1.0.1
plotly==6.0.1
fastapi==0.115.12
uvicorn==0.34.0
scikit-learn==1.6.1
joblib==1.4.2
lightgbm==4.6.0
catboost==1.2.8
stable-baselines3==2.6.0
gymnasium==1.1.1
py-clob-client
websockets==15.0.1
